+++ R/checks.R
@@ -65,13 +65,14 @@
 }
 
 .check_input_rows <- function(x, y) {
-  if (nrow(x) != nrow(y))
+  if (nrow(x) != nrow(y)) {
     stop(
       sprintf(
         "'%s' must have the same number of rows as '%s'.",
         deparse(substitute(x)), deparse(substitute(y))
       )
     )
+  }
 }
 
 .check_bbox <- function(bbox) {
@@ -130,7 +131,7 @@
   modes <- c("fast", "short")
   if (!routing_mode %in% modes) {
     stop(
-        sprintf(
+      sprintf(
         "Routing mode '%s' not valid, must be in ('%s').",
         routing_mode,
         paste(modes, collapse = "', '")
@@ -153,7 +154,7 @@
   if (!(is.character(api_key) & api_key != "")) {
     stop(
       "Please provide an 'API key' for a HERE project.
-         Get your login here: https://developer.here.com/"
+      Get your login here: https://developer.here.com/"
     )
   }
 }