+++ R/connection.R
@@ -72,7 +72,7 @@ connection <- function(origin, destination, datetime = Sys.time(),
   )
 
   # Add departure and arrival
-  url = paste0(
+  url <- paste0(
     url,
     "&origin=",
     coords_orig,
@@ -104,19 +104,23 @@ connection <- function(origin, destination, datetime = Sys.time(),
   }
 
   # Add route attributes
-  url = paste0(
+  url <- paste0(
     url,
     "&return=polyline,travelSummary"
   )
 
   # Return urls if chosen
-  if (url_only) return(url)
+  if (url_only) {
+    return(url)
+  }
 
   # Request and get content
   data <- .get_content(
     url = url
   )
-  if (length(data) == 0) return(NULL)
+  if (length(data) == 0) {
+    return(NULL)
+  }
 
   # Extract information
   routes <- .extract_connection_sections(data)
@@ -138,10 +142,14 @@ connection <- function(origin, destination, datetime = Sys.time(),
     routes <- .connection_summary(routes)
   }
 
+  # Bug of data.table and sf combination? Drops sfc class, when only one row...
+  routes <- as.data.frame(routes)
+  routes$geometry <- sf::st_sfc(routes$geometry, crs = 4326)
+
   # Create sf object
   return(
     sf::st_as_sf(
-      as.data.frame(routes),
+      routes,
       sf_column_name = "geometry",
       crs = 4326
     )
@@ -156,6 +164,7 @@ connection <- function(origin, destination, datetime = Sys.time(),
   template <- data.table::data.table(
     id = numeric(),
     rank = numeric(),
+    section = numeric(),
     departure = character(),
     origin = character(),
     arrival = character(),
@@ -170,13 +179,16 @@ connection <- function(origin, destination, datetime = Sys.time(),
     geometry = character()
   )
   routes <- data.table::rbindlist(
-    append(list(template),
+    append(
+      list(template),
       lapply(data, function(con) {
         count <<- count + 1
 
         # Parse JSON
         df <- jsonlite::fromJSON(con)
-        if (is.null(df$routes$sections)) {return(NULL)}
+        if (is.null(df$routes$sections)) {
+          return(NULL)
+        }
 
         # Connections
         rank <- 0
@@ -189,14 +201,19 @@ connection <- function(origin, destination, datetime = Sys.time(),
               rank <<- rank + 1
               data.table::data.table(
                 rank = rank,
+                section = seq_len(nrow(sec)),
                 departure = sec$departure$time,
-                origin = vapply(sec$departure$place$name,
-                                function(x) if (is.na(x)) "ORIG" else x,
-                                character(1)),
+                origin = vapply(
+                  sec$departure$place$name,
+                  function(x) if (is.na(x)) "ORIG" else x,
+                  character(1)
+                ),
                 arrival = sec$arrival$time,
-                destination = vapply(sec$arrival$place$name,
-                                     function(x) if (is.na(x)) "DEST" else x,
-                                     character(1)),
+                destination = vapply(
+                  sec$arrival$place$name,
+                  function(x) if (is.na(x)) "DEST" else x,
+                  character(1)
+                ),
                 mode = sec$transport$mode,
                 category = sec$transport$category,
                 vehicle = sec$transport$name,
@@ -206,13 +223,19 @@ connection <- function(origin, destination, datetime = Sys.time(),
                 duration = sec$travelSummary$duration,
                 geometry = sec$polyline
               )
-            }), fill = TRUE)
+            }),
+            fill = TRUE
+          )
         )
-      })), fill = TRUE
-    )
+      })
+    ),
+    fill = TRUE
+  )
 
   # Check success
-  if (nrow(routes) < 1) {return(NULL)}
+  if (nrow(routes) < 1) {
+    return(NULL)
+  }
 
   # Decode flexible polyline encoding to LINESTRING
   routes$geometry <- sf::st_geometry(
@@ -230,7 +253,7 @@ connection <- function(origin, destination, datetime = Sys.time(),
     departure = min(departure),
     origin = origin[2],
     arrival = max(arrival),
-    destination = destination[length(destination)-1],
+    destination = destination[length(destination) - 1],
     transfers = length(stats::na.exclude(vehicle)) - 1,
     modes = paste(stats::na.exclude(mode), collapse = ", "),
     categories = paste(stats::na.exclude(category), collapse = ", "),
@@ -238,8 +261,7 @@ connection <- function(origin, destination, datetime = Sys.time(),
     providers = paste(stats::na.exclude(provider), collapse = ", "),
     distance = sum(distance),
     duration = sum(duration),
-    geometry = sf::st_union(geometry)
+    geometry = suppressMessages(sf::st_union(geometry))
   ), by = list(id, rank)]
   return(summary)
 }
-