+++ NEWS.md
@@ -1,3 +1,12 @@
+# version 0.6.1
+
+* Use **styler** package and use `tyler::tidyverse_style()`to format the package.
+* All **lintr** issues are solved, except from line length issues (limit of 80 characters).
+* Avoid drop of the `sfc` class of the geometry column when there is only one geometry/row in a `data.table` (closes [#111](https://github.com/munterfinger/hereR/issues/111)).
+* Return `queryScore` in `geocode()` results as `score` column (closes [#109](https://github.com/munterfinger/hereR/issues/109)).
+* Silence **sf** messages of `sf::st_union` call in `connection(..., summary = TRUE)`.
+* Added section id to `route()`, `intermodal_route()` and `connection()`.
+
 # version 0.6.0
 
 * Update `route_matrix()` from **Routing API v7.2** (calculatematrix) to **Matrix Routing API v8** (see [#87](https://github.com/munterfinger/hereR/issues/87)).